fix: Decimal precision, dedup backtest engine, add LiveBroker portfolio tracking
- Strategy._get_closes now returns list[Decimal] instead of list[float],
  preserving price precision at the interface boundary. MA Crossover
  and RSI Reversal strategies updated to work with Decimal internally.
- BacktestEngine._record_equity and _update_mark_price now delegate to
  Portfolio.recompute_equity() and Portfolio.update_mark_prices(),
  removing duplicated PnL/drawdown logic. Removed dead _get_bar_at.
- LiveBroker._apply_fill syncs portfolio positions when orders fill,
  matching PaperBroker behavior — live trading portfolio was
  previously never updated on fills.

+++ crypto_trading/core/strategy.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 from collections import defaultdict
+from decimal import Decimal
 from typing import Any
 
 from crypto_trading.core.types import OHLCV, Signal, Ticker
@@ -35,5 +36,5 @@ class Strategy(ABC):
     def _get_bars(self, symbol: str, count: int) -> list[OHLCV]:
         return self._bar_history[symbol][-count:]
 
-    def _get_closes(self, symbol: str, count: int) -> list[float]:
-        return [float(b.close) for b in self._bar_history[symbol][-count:]]
+    def _get_closes(self, symbol: str, count: int) -> list[Decimal]:
+        return [b.close for b in self._bar_history[symbol][-count:]]
